requirements.txt
llama-cpp-python
pyyaml
huggingface_hub
rich